chịu

+++ src/main/java/com/orfarmweb/repository/ProductRepo.java
@@ -12,7 +12,7 @@ public interface ProductRepo extends JpaRepository<Product, Integer> {
     List<Product> findProductByCategoryId(int id);
     @Query(value = "select * from product left join category on product.cate_id = category.id " +
             "where product.sale_price between ?1 and ?2 and category.id = ?3", nativeQuery = true)
-    List<Product> listFill(int a, int b, int id);
+    List<Product> listFill(float a, float b, int id);
     @Query(value = "select count(product.id) from product left join category on product.cate_id = category.id where category.id = ?",nativeQuery = true)
     int getTotal(int id);
     @Query(value = "select * from product where id = ?", nativeQuery = true)
